♻️ refactor: 재귀함수 /util 로 이동
