+++ app/(dashboard)/mydashboard/_components/dashboard-list.tsx
@@ -3,55 +3,11 @@
 import PageButton from '@/app/components/pagination/page-button';
 import { TEAM_BASE_URL } from '@/constants/TEAM_BASE_URL';
 import { DashboardDetail } from '@/types/dashboard';
+import makeDashboardArr from '@/utils/makeDashboardResponse';
 import { getCookie } from 'cookies-next';
 import { useEffect, useState } from 'react';
 
 import DashboardCard from './dashboard-card';
-
-const token = getCookie('token');
-
-// NOTE: 대시보드 중복 제거 후 6개로 맞추는 함수
-async function makeDashboardArr(
-  arr: DashboardDetail[],
-  page: number,
-  size: number
-): Promise<DashboardDetail[]> {
-  const url = `${TEAM_BASE_URL}/dashboards?navigationMethod=pagination&page=${page + 1}&size=${size}`;
-
-  async function reFetch() {
-    const res = await fetch(url, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${token}`,
-      },
-    });
-
-    const data = await res.json();
-    const { dashboards } = data;
-    return dashboards;
-  }
-
-  const fetchArr: DashboardDetail[] = await reFetch();
-
-  const checkDashboard: DashboardDetail[] = fetchArr.reduce(
-    (acc: DashboardDetail[], cur: DashboardDetail) => {
-      if (acc.findIndex(({ id }) => id === cur.id) === -1) {
-        acc.push(cur);
-      }
-      return acc;
-    },
-    []
-  );
-
-  arr.push(...checkDashboard);
-
-  if (arr.length < 6) {
-    makeDashboardArr(arr, page + 1, size - arr.length);
-  }
-
-  return arr;
-}
 
 interface DashboardListProps {
   initialData: DashboardDetail[];
@@ -65,6 +21,8 @@ export default function DashboardList({
   const [page, setPage] = useState(1);
   const [dashboardList, setDashboardList] =
     useState<DashboardDetail[]>(initialData);
+
+  const token = getCookie('token');
 
   async function getData() {
     const url = `${TEAM_BASE_URL}/dashboards?navigationMethod=pagination&page=${page}&size=6`;
@@ -93,7 +51,12 @@ export default function DashboardList({
     // NOTE: 6개 이하일 경우(=중복 데이터 존재) 함수 호출
     if (checkDashboard.length < 6) {
       setDashboardList(
-        await makeDashboardArr(checkDashboard, page, 6 - checkDashboard.length)
+        await makeDashboardArr(
+          checkDashboard,
+          page,
+          6 - checkDashboard.length,
+          token
+        )
       );
       return;
     }
